pip_requirements.txt
BeautifulSoup==3.2.1
Django==1.4.3
PIL==1.1.7
South==0.7.6
anyjson==0.3.3
billiard==2.7.3.19
django-admin-bootstrapped==0.3.2
django-extensions==1.0.2
django-haystack==1.2.7
django-i18nurls==0.7
django-tagging==0.3.1
docutils==0.10
python-dateutil==1.5
pytz==2012j
readline==6.2.4.1
xapian-haystack==1.1.5beta
